requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.7
Pillow==10.0.1
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
